refactor: BigDecimal -> double + user opinions + top tags

+++ src/main/java/com/truskappka/truskappka_backend/common/dto/Coordinate.java
@@ -1,8 +1,6 @@
 package com.truskappka.truskappka_backend.common.dto;
 
-import java.math.BigDecimal;
-
 public record Coordinate(
-        BigDecimal longitude,
-        BigDecimal latitude
+        double longitude,
+        double latitude
 ) {}
