+++ src/main/java/com/truskappka/truskappka_backend/opinion/dto/Rating.java
@@ -1,9 +1,7 @@
 package com.truskappka.truskappka_backend.opinion.dto;
 
-import java.math.BigDecimal;
-
 public record Rating(
-        BigDecimal quality,
-        BigDecimal service,
-        BigDecimal price
+        Integer quality,
+        Integer service,
+        Integer price
 ) {}
